Split presenter package up into modules
Move AnnotationHTMLPresenter and DocumentHTMLPresenter into their own
independently tested modules. This is preparation for moving the memex
presenters into h.
